requirements.txt
requests==2.31.0
openai==1.3.7
RPLCD==1.3.1
smbus2==0.4.2
python-dotenv==1.0.0